- Refactor user-role hierarchy: remove Manager, add Admin and SuperUser tiers, and update authorization checks. - Add migration scripts to map legacy Admin users to SuperUser and legacy SuperUser/Manager users to SchoolAdmin. - Set MFA code expiry to 30 minutes by default and 8 hours for Admin and SuperUser users. - Add MFA regression tests for normal, Admin, and SuperUser login flows. - Update enrollment and attendance email templates to use “Assalaamu alaikum” and participant-based wording. - Replace cancellation email text with:   “The participant's registration has been cancelled.”   “L'inscription du participant a été annulée.” - Add email-template regression coverage.

+++ Application.Users.Repository.Implementation/TempUserRepository.cs
@@ -37,7 +37,7 @@ namespace Application.Users.Repository.Implementation
                     cmd.AddParameter("@createdAt", userInformation.CreatedAt);
                     cmd.AddParameter("@updatedOn", userInformation.CreatedAt);
                     cmd.AddParameter("@relationship", (int)userInformation.Relationship);
-                    cmd.AddParameter("@userRole", (int)userInformation.UserRole);
+                    cmd.AddParameter("@userRole", (long)userInformation.UserRole);
                     
                     if (await cmd.ExecuteNonQueryAsync().ConfigureAwait(false) > 0)
                     {
@@ -162,7 +162,7 @@ namespace Application.Users.Repository.Implementation
                     var isTempPassword = reader.GetBoolean(10);
                     var relationship = (Relationship)reader.GetInt32(11);
                     var familyId = reader.GetGuidFromByteArray(12);
-                    var userRole = (UserRoleType)reader.GetInt32(13);
+                    var userRole = (UserRoleType)reader.GetInt64(13);
 
                     return new UserInformation
                     {
@@ -247,7 +247,7 @@ namespace Application.Users.Repository.Implementation
                     cmd.AddParameter("@updatedOn", DateTime.UtcNow);
                     cmd.AddParameter("@isTempPassword", userInformation.IsTempPassword);
                     cmd.AddParameter("@relationship", (int)userInformation.Relationship);
-                    cmd.AddParameter("@userRole", (int)userInformation.UserRole);
+                    cmd.AddParameter("@userRole", (long)userInformation.UserRole);
 
                     if (await cmd.ExecuteNonQueryAsync().ConfigureAwait(false) > 0)
                     {
@@ -290,7 +290,7 @@ namespace Application.Users.Repository.Implementation
                         var UpdatedOn = reader.GetDateTime(9);
                         var isTempPassword = reader.GetBoolean(10);
                         var relationship = (Relationship)reader.GetInt32(11);
-                        var userRole = (UserRoleType)reader.GetInt32(12);
+                        var userRole = (UserRoleType)reader.GetInt64(12);
                         var familyId = reader.GetGuidFromByteArray(13);
 
                         results.Add(new UserInformation
@@ -350,7 +350,7 @@ namespace Application.Users.Repository.Implementation
                         var updatedOn = reader.GetDateTime(9);
                         var isTempPassword = reader.GetBoolean(10);
                         var relationship = (Relationship)reader.GetInt32(11);
-                        var userRole = (UserRoleType)reader.GetInt32(12);
+                        var userRole = (UserRoleType)reader.GetInt64(12);
                         var tempFamilyId = reader.GetGuidFromByteArray(13);
 
                         results.Add(new UserInformation
@@ -417,7 +417,7 @@ namespace Application.Users.Repository.Implementation
                     var isTempPassword = reader.GetBoolean(7);
                     var relationship = (Relationship)reader.GetInt32(8);
                     var familyId = reader.GetGuidFromByteArray(9);
-                    var userRole = (UserRoleType)reader.GetInt32(10);
+                    var userRole = (UserRoleType)reader.GetInt64(10);
 
                     return new UserInformation
                     {
